+++ src/editor/config.js
@@ -79,6 +79,7 @@ export const CLASS_OR_ID = genUpper2LowerKeyHash([
   'AG_TASK_LIST',
   'AG_TASK_LIST_ITEM',
   'AG_TASK_LIST_ITEM_CHECKBOX',
+  'AG_CHECKBOX_CHECKED',
   'AG_TABLE_TOOL_BAR'
 ])