+++ src/editor/contentState/updateCtrl.js
@@ -138,9 +138,9 @@ const updateCtrl = ContentState => {
   // handle task list item checkbox click
   ContentState.prototype.listItemCheckBoxClick = function (checkbox) {
     const { checked, id } = checkbox
-    console.log(checked)
     const block = this.getBlock(id)
     block.checked = checked
+    this.render()
   }
 
   ContentState.prototype.updateList = function (block, type, marker = '') {