+++ src/renderer/components/search.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="search-bar">
+
+  </div>
+</template>
+
+<script>
+  
+</script>
+
+<style scoped>
+  
+</style>