+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - fix the bug that click at the edge of code block will caused the code block does not be focused.
 
+**Optimization**
+
+- Optimize the display of word count in title bar. we also delete the background color of title bar to make it more concise.
+
+- Customize the style of checkbox in Task List Item.
+
+- Change the display of Insert Table dialog.
+
 ### 0.3.0
 
 **Features**